Add Windows build test (#68657)
Enables build_test for Windows. Currently only flutter_gallery has platform directories for the desktop platforms, so this will run only that build, but this will provide an end-to-end build test for Windows.

Other example/test projects can be brought online for Windows in the future just by adding the windows/ directory to the project.