[flutter_tools] use temp dir for tester dill, use desktop log reader (#75922)
On windows, its not possible to run 2 integration tests using flutter-tester simultaneously. This is due to the flutter tester using a fixed location for its dill output. Use a temp dir instead.

Also replace the custom log reader in flutter-tester with the desktop log reader, and ensure stopApp waits for the process exit code.

Fixes #75534