Skip the iteration in Layer._fireCompositionCallbacks if the callbacks map is empty (#130438)
This was showing up as a hot spot in some benchmarks and profiles. This
function is called frequently during frame builds and often has an empty
map. There may be significant overhead from obtaining the values
iterator and cloning it into a list.

See https://github.com/flutter/flutter/issues/130339